pyproject.toml
[project]
name = "jsm-asset-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]
